Готовая 3 лаба

+++ Алгоритмы и теория сложности/Task3/src/CalcListener.java
@@ -1,4 +1,4 @@
 public interface CalcListener {
-    void stateChanged(double[] mass1, double[] mass2, double[] mass3);
+    void stateChanged(double[] mass1, int[] mass2, int[] mass3);
     void setResult(String res);
 }
